AttentionCodes
- Cleaned up some unnecessary overrides
- Some var name changes
- Starting work on AttantionCode

AttantionCode will be a byte of bit flags, same var can be used for
either city, ship, or other. Depending on the type of object it will
then mean diffenent things, the first flags I added to the city object
is "jobs > homes",  "population < homes", and "morale < 20" warning.
Will be adding flags for things like:
- City: No/Low food
- City: Enermy detected
- Ship: No/Low fuel
- Ship: Damaged
- Ship: Enermy detected