Mail reading error check
Added a second ProgressBar, so there is now one for reading files and
one for adding them to the tables, this make it easier to see when/where
a crash happened.

Added a Try Catch to file reading, if an exception is thrown it will
popup and tell the user what mail file caused the crash and give an
option for adding the file path to the clipboard.

Fixes: https://github.com/Deantwo/HazeronAdviser/issues/7