WIP Over Population
Underlayinng work on getting the Population Limit of a cities planet.
Not totally sure if this is even a good idea, because there can be more
than one city in a resource zone.
But still gonna need it for when I start my work on the graphs!

I do however seem to have a problem with not knowing what the population
limit is when a planet is inbetween two of the rages.