Mining for Names
Wanting to add names to the character filter, I created an object for
character ID and name comparison.
When scanning mails the program now checks all MSG_Mail type mails and
saves the sender's ID and Name.
This was the fastest and easiest way to get the names of characters
because the recipient's name is never displayed in a mail.

Also:
- Minor lambda optimisation
- Using the int value of character IDs for comparison rather than string